+++ src/main/java/naem/server/exception/CustomException.java
@@ -1,4 +1,4 @@
-package naem.server.exhandler;
+package naem.server.exception;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
